rag_mercadolibre/requirements.txt
streamlit
langchain==0.2.11
langchain-google-genai
google-genai
langchain-weaviate
langchain-core
selenium
python-dotenv
pandas
openpyxl
weaviate-client>=4.6.0
langchain-text-splitters
langchain-community
PyPDF2==3.0.1
python-multipart==0.0.6